[refactor] Move version check logic to kWordleApp.swift
