+++ kWordle/ViewModel/MainViewModel.swift
@@ -27,13 +27,12 @@ class MainViewModel: ObservableObject {
     
     @Published var isResultAnimationPlaying: Bool = false
     @Published var isInvalidWordWarningPresented: Bool = false
-
-    @Published var needUpdate: Bool = false
     
     @Published var isADNotLoaded: Bool = false
     
     @AppStorage("life") var life = UserDefaults.standard.integer(forKey: "life")
     @AppStorage("lifeTimeStamp") var lifeTimeStamp: String = UserDefaults.standard.string(forKey: "lifeTimeStamp") ?? ""
+    
     let rewardADViewController = RewardedADViewController()
 
     var preventTapShowAdButton: Bool = false
@@ -47,10 +46,6 @@ class MainViewModel: ObservableObject {
         hintRow = Self.getEmptyHintRow(length: 5)
         markCorrectJamoOnHint()
         lifeCount = life
-//        checkUpdate { updateNeeded in
-//            DispatchQueue.main.async {
-//                self.needUpdate = updateNeeded
-//            }}
         checkLifeCount()
         print(game.answer)
     }
@@ -132,52 +127,6 @@ class MainViewModel: ObservableObject {
         let deivceUUID = UIDevice.current.identifierForVendor?.uuidString ?? ""
         Analytics.logEvent(state + "-" + deivceUUID, parameters: [:])
     }
-    
-    private func checkUpdate(completion: @escaping (Bool) -> Void) {
-        let appVersion = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String
-        latestVersion { latestVersion in
-            guard let latestVersion = latestVersion else { completion(false); return }
-            if let lastCheckVersion = UserDefaults.standard.string(forKey: "latestVersion") {
-                let compareVersion = lastCheckVersion.compare(latestVersion, options: .numeric)
-                if compareVersion != .orderedAscending {
-                    completion(false); return
-                }
-            }
-            UserDefaults.standard.set(latestVersion, forKey: "latestVersion")
-            
-            let compareResult = appVersion?.compare(latestVersion, options: .numeric)
-            
-            if compareResult == .orderedAscending {
-                completion(true)
-            } else {
-                completion(false)
-            }
-        }
-    }
-    
-    func latestVersion(completion: @escaping (String?) -> Void) {
-        let appleID = "1619947572"
-        guard let url = URL(string: "https://itunes.apple.com/lookup?id=\(appleID)") else {
-            completion(nil); return
-        }
-        let task = URLSession.shared.dataTask(with: url) { (data, _, _) in
-            guard let data = data,
-                  let json = try? JSONSerialization.jsonObject(with: data, options: .allowFragments) as? [String: Any],
-                  let results = json["results"] as? [[String: Any]],
-                  let appStoreVersion = results[0]["version"] as? String else {
-                completion(nil); return
-            }
-            completion(appStoreVersion)
-        }
-        task.resume()
-    }
-    
-    func openAppStore() {
-        let appleID = "1619947572"
-        let appStoreOpneUrlString = "itms-apps://itunes.apple.com/app/apple-store/\(appleID)"
-        guard let url = URL(string: appStoreOpneUrlString) else { return }
-        UIApplication.shared.open(url, options: [:], completionHandler: nil)
-    }
 }
 
 /// Ad
@@ -202,15 +151,11 @@ extension MainViewModel {
     }
     
     func startNewGameWithAd() {
-        presentAdFor {
-            self.startNewGame()
-        }
+        presentAdFor { self.startNewGame() }
     }
     
     func addLifeCountWithAD() {
-        presentAdFor {
-            self.addLifeCount()
-        }
+        presentAdFor { self.addLifeCount() }
     }
     
     func showHintWithAd(revealHintAt index: Int) {
